+++ app/src/main/AndroidManifest.xml
@@ -16,7 +16,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.AppCompat.Light.NoActionBar">
-        <activity android:name=".PastProjectActivity"></activity>
+        <activity android:name=".activities.PastProjectActivity"></activity>
         <activity android:name=".DashboardPhase1Activity" />
         <activity android:name=".DashboardPhase2Activity" />
         <activity android:name=".DashboardPhase3Activity" />
@@ -42,7 +42,7 @@
         </activity>
 
         <activity android:name=".activities.MainActivity" />
-        <activity android:name=".ChatActivity"
+        <activity android:name=".activities.ChatActivity"
             android:theme="@style/AppTheme"></activity>
 
         <activity android:name=".activities.GuruProfileActivity"/>
