+++ app/src/main/java/com/bluesky/toa/DashboardPhase3Activity.java
@@ -6,7 +6,9 @@ import android.view.View;
 import android.widget.Button;
 
 import com.bluesky.toa.activities.DashBoardActivity;
+import com.bluesky.toa.activities.GuruProfileActivity;
 import com.bluesky.toa.activities.HomePageActivity;
+import com.bluesky.toa.activities.PastProjectActivity;
 
 import androidx.appcompat.app.AppCompatActivity;
 
@@ -32,7 +34,7 @@ public class DashboardPhase3Activity extends AppCompatActivity {
 
         projectDasboardBtnEPF.setOnClickListener(new View.OnClickListener(){
             public void onClick(View v){
-                Intent intent = new Intent(DashboardPhase3Activity.this, HomePageActivity.class);
+                Intent intent = new Intent(DashboardPhase3Activity.this, GuruProfileActivity.class);
                 startActivity(intent);
             }
         });
