+++ app/src/main/java/com/bluesky/toa/activities/PastProjectActivity.java
@@ -1,4 +1,4 @@
-package com.bluesky.toa;
+package com.bluesky.toa.activities;
 
 import androidx.appcompat.app.AppCompatActivity;
 
@@ -7,8 +7,7 @@ import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
 
-import com.bluesky.toa.activities.DashBoardActivity;
-import com.bluesky.toa.activities.HomePageActivity;
+import com.bluesky.toa.R;
 
 public class PastProjectActivity extends AppCompatActivity {
 
@@ -23,7 +22,7 @@ public class PastProjectActivity extends AppCompatActivity {
 
         projectDasboardBtnEPF.setOnClickListener(new View.OnClickListener(){
             public void onClick(View v){
-                Intent intent = new Intent(PastProjectActivity.this, HomePageActivity.class);
+                Intent intent = new Intent(PastProjectActivity.this, GuruProfileActivity.class);
                 startActivity(intent);
             }
         });
