+++ app/src/main/res/layout/activity_past_project.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".PastProjectActivity">
+    tools:context=".activities.PastProjectActivity">
 
     <TextView
         android:id="@+id/dash_board_pp_title_jj"
